Goodbye Dagger in data:api module
