+++ data/api/build.gradle
@@ -33,14 +33,22 @@ kotlin {
                 implementation Dep.Kotlin.serializationAndroid
                 implementation Dep.OkHttp.client
                 implementation Dep.Stetho.stetho
-                implementation Dep.Dagger.core
-                implementation Dep.Dagger.androidSupport
+                implementation Dep.DependencyProperty.dependencyProperty
                 implementation Dep.OkHttp.loggingInterceptor
             }
         }
         androidTest {
             dependsOn commonMain
             dependencies {
+                implementation Dep.Kotlin.stdlibJvm
+                implementation Dep.Kotlin.androidCoroutinesDispatcher
+                implementation Dep.Ktor.clientAndroid
+                implementation Dep.Ktor.serializationJvm
+                implementation Dep.Kotlin.serializationAndroid
+                implementation Dep.OkHttp.client
+                implementation Dep.Stetho.stetho
+                implementation Dep.DependencyProperty.dependencyProperty
+                implementation Dep.OkHttp.loggingInterceptor
                 implementation Dep.Test.junit
                 implementation Dep.Test.slf4j
             }
@@ -62,6 +70,5 @@ android {
 }
 
 dependencies {
-    kapt Dep.Dagger.compiler
-    kapt Dep.Dagger.androidProcessor
+    implementation Dep.DependencyProperty.dependencyProperty
 }
