+++ data/api/src/main/java/io/github/droidkaigi/confsched2020/data/api/internal/InjectableKtorGoogleFormApi.kt
@@ -1,8 +1,7 @@
 package io.github.droidkaigi.confsched2020.data.api.internal
 
 import io.ktor.client.HttpClient
-import javax.inject.Inject
 
-internal class InjectableKtorGoogleFormApi @Inject constructor(
+internal class InjectableKtorGoogleFormApi(
     httpClient: HttpClient
 ) : KtorGoogleFormApi(httpClient)
